requirements.txt
torch>=2.0.0
transformers>=4.40.0
peft>=0.10.0
datasets>=2.18.0
accelerate>=0.28.0
openai>=1.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
jsonlines>=4.0.0
scikit-learn>=1.4.0
